+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes from v1.0.0 will be documented in this file.
 
+## [1.4.0] - 2019-11-09
+### Changed
+- Inputs of Map
+- Inputs of Placemark
+- Inputs of MultiRoute
+- Inputs in README
+
 ## [1.3.0] - 2019-10-30
 ### Added
 - Angular 6+ support
@@ -36,6 +43,7 @@ All notable changes from v1.0.0 will be documented in this file.
 ## [1.0.0] - 2019-09-14
 - Release without changes
 
+[1.4.0]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/ddubrava/angular-yandex-maps/compare/v1.1.0...v1.2.0